+++ app/Observers/ProjectObserver.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Observers;
 
 use App\Models\Project;
@@ -21,11 +23,9 @@ class ProjectObserver
     public function updated(Project $project): void
     {
         // check if image is updated
-        if ($project->isDirty('main_image')) {
-            // delete old image
-            if (Storage::disk('public')->exists($project->getOriginal('main_image'))) {
-                Storage::disk('public')->delete($project->getOriginal('main_image'));
-            }
+        // delete old image
+        if ($project->isDirty('main_image') && Storage::disk('public')->exists($project->getOriginal('main_image'))) {
+            Storage::disk('public')->delete($project->getOriginal('main_image'));
         }
 
         // check if galleries are updated
@@ -66,12 +66,10 @@ class ProjectObserver
             Storage::disk('public')->delete($project->main_image);
         }
         // delete galleries
-        if (count($project->galleries) > 0) {
-            // delete galleries
-            foreach ($project->galleries as $gallery) {
-                if (Storage::disk('public')->exists($gallery)) {
-                    Storage::disk('public')->delete($gallery);
-                }
+        // delete galleries
+        foreach ($project->galleries as $gallery) {
+            if (Storage::disk('public')->exists($gallery)) {
+                Storage::disk('public')->delete($gallery);
             }
         }
     }
